tests/wpt/tests/mathml/relations/css-styling/parsing/math-style-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>MathML Core: parsing math-style with valid values</title>
<link rel="author" title="Eri Pazos" href="mailto:eri@igalia.com">
<link rel="help" href="https://w3c.github.io/mathml-core/#the-math-style-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
  test_valid_value("math-style", "normal");
  test_valid_value("math-style", "compact");
</script>
</body>
</html>
